work in progress in gpm refactoring
