+++ gpm_api_consumer/__init__.py
@@ -1,9 +1,10 @@
 from .core.Operators import GPMOperator
 from .utils import decorators
 from .core import exceptions as gpm_exceptions
-
+from .core.Middleware import MiddleWare
 __all__ = [
     "GPMOperator",
     "decorators",
     "gpm_exceptions",
+    "MiddleWare",
 ]
