+++ gpm_api_consumer/utils/__init__.py
@@ -1,5 +1,7 @@
-from .utils import normalize_name
+from .utils import normalize_name, set_logger_level, chunked_iterable
 
 __all__ = [
     "normalize_name",
+    "set_logger_level",
+    "chunked_iterable",
 ]
